pyproject.toml
[project]
name = "archiveflow"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
]

[dependency-groups]
dev = [
    "flake8>=7.1.1",
    "ipython>=8.18.1",
    "isort>=5.13.2",
    "mypy>=1.13.0",
    "pytest>=8.3.4",
    "ruff>=0.8.3",
    "types-requests>=2.32.0.20241016",
]
